refactor: create a new page component from page.tsx
